ビジネスロジックチックなところをmodelに移動
