+++ router/router.go
@@ -16,8 +16,8 @@ func SetUpRouter() *gin.Engine {
 		auth.POST("/register", controllers.Register)
 		auth.POST("/login", controllers.Login)
 	}
-	public.GET("/organizations-ranking",controllers.GetOrganizationsRanking)
-	
+	public.GET("/organizations-result",controllers.GetOrganizationsRanking)
+
 	protected := router.Group("/api")
 	protected.Use(middlewares.JwtAuthMiddleware())
 	return router
